% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_time_dep.R
\name{time_dep}
\alias{time_dep}
\title{Create a time-dependent variable}
\usage{
time_dep(time, value, check = TRUE)
}
\arguments{
\item{time}{Numeric vector. Must contain \code{-Inf}, and may include \code{Inf}}

\item{value}{Values taken at various times. the value at \code{-Inf} is the value
of the variable before any finite time.}

\item{check}{If \code{TRUE}, checks whether the input is valid.}
}
\value{
A data frame of class \code{time_dep}.
}
\description{
Given a vector of times (ordered), and values (factor or numeric), forms a
"time-dependent" data frame. The value of the variable changes at time s to
the value x corresponding to s. In other words, x is the value of the
variable on (s, t], where t is the next time the variable changes, possibly
Inf
}
\examples{
x <- time_dep(c(5.1, -Inf), c(1, 0))
x
}
